+++ pub/config.go
@@ -13,7 +13,7 @@ type Config struct {
 	LifecycleUrl          string `env:"LIFECYCLE_URL" envDefault:"http://localhost:7202/lifecycle"`
 	AuthRequired          bool   `env:"AUTH_REQUIRED" envDefault:"false"`
 	AuthDebug             bool   `env:"AUTH_DEBUG" envDefault:"false"`
-	LifecycleToken        string `env:"LIFECYCLE_TOKEN"` // Used to authenticate to Lifecycle for User endpoints
+	LifecycleToken        string `env:"LIFECYCLE_AUTH_TOKEN"` // Used to authenticate to Lifecycle for User endpoints
 	RequestTracingHeader  string `env:"REQUEST_TRACING_HEADER" envDefault:"X-Request-Tracing-Id"`
 	OpenTelemetryEndpoint string `env:"OPENTELEMETRY_ENDPOINT" envDefault:""`
 }