+++ racetrack_commons/racetrack_commons/auth/token.py
@@ -72,7 +72,7 @@ def generate_service_token(
     auth_secret_key = os.environ.get('AUTH_KEY')
     assert auth_secret_key, 'fill AUTH_KEY var with your secret key'
     token = generate_internal_token(subject_name, auth_secret_key, scopes=[AuthScope.FULL_ACCESS])
-    print(f'Service: {subject_name}\nLIFECYCLE_TOKEN:\n{token}\n')
+    print(f'Service: {subject_name}\nLIFECYCLE_AUTH_TOKEN:\n{token}\n')
 
 
 def generate_internal_token(