+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Deployment phases, shown by a racetrack client, are more granular,
+  including image building steps.
 
 ## [2.8.1] - 2023-01-27
 ### Changed